Add files via upload
Working LED test allows the user to send a signal, in the form of a button press or a switch or a potentiometer or whatever, to another radio device to turn on an LED. Allows for communication in either direction but only one way at a time. Easy serial interface to change direction implemented.