Fixed the user information to the serial port.
Cleaned up the UI. Usually works, but often times the first and second writes both from the master and sensor nodes do not return true. Not sure what the problem is but I may try to write in a while loop with a break condition of sending correctly or not being connected to the network.